Restructure project
- Move Node client to /node
- Move Go client (and server) under /pkg
- Move Node scenario tests to /scenario/node
- Move Go scenario tests to /scenario/go
- Remove prototype client applications
- Move wallet code out of /pkg into /prototype as current implementation does not use wallets
- Better definition of Node packages (client and scenario)
- Refactor Makefile for new structure and to tighten up dependencies
- Use correct CouchDB docker image versions for Fabric 2.2
