--- client/go/client2.go
@@ -1,67 +1,0 @@
-/*
-Copyright 2020 IBM All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package main
-
-import (
-	"flag"
-	"fmt"
-	"log"
-
-	"github.com/hyperledger/fabric-gateway/client/go/sdk"
-	"github.com/hyperledger/fabric-gateway/pkg/identity"
-	"github.com/hyperledger/fabric-gateway/pkg/wallet"
-)
-
-func main() {
-	idPath := flag.String("id", "", "path to the client's wallet identity")
-	flag.Parse()
-
-	walletIdentity, err := wallet.ReadWalletIdentity(*idPath)
-	if err != nil {
-		log.Fatalf("failed to read client identity: %s", err)
-	}
-
-	certificate, err := identity.CertificateFromPEM([]byte(walletIdentity.Credentials.Certificate))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	id, err := identity.NewX509Identity(walletIdentity.MspID, certificate)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	privateKey, err := identity.PrivateKeyFromPEM([]byte(walletIdentity.Credentials.Key))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	signer, err := identity.NewPrivateKeySign(privateKey)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	gw, err := sdk.Connect("localhost:1234", id, signer)
-	defer gw.Close()
-
-	network := gw.GetNetwork("mychannel")
-	contract := network.GetContract("fabcar")
-
-	try(contract.Evaluate("queryAllCars"))
-	try(contract.SubmitSync("createCar", sdk.WithArguments("CAR10", "VW", "Polo", "Grey", "Mary")))
-	try(contract.Evaluate("queryCar", sdk.WithArguments("CAR10")))
-	try(contract.SubmitSync("changeCarOwner", sdk.WithArguments("CAR10", "Archie")))
-	try(contract.Evaluate("queryCar", sdk.WithArguments("CAR10")))
-
-}
-
-func try(result []byte, err error) {
-	if err != nil {
-		log.Panicf("Error: %s", err)
-	}
-	fmt.Println(string(result))
-}
