+++ node/package.json
@@ -3,16 +3,17 @@
   "version": "0.0.1",
   "description": "Node SDK client library for Hyperledger Fabric Gateway",
   "main": "sdk.js",
+  "engines": {
+    "node": "^10.13.0 || ^12.13.0"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo 'No unit tests (yet)'"
   },
   "author": "",
-  "license": "ISC",
+  "license": "Apache-2.0",
   "dependencies": {
     "@grpc/grpc-js": "^1.1.5",
     "@grpc/proto-loader": "^0.5.5",
-    "chai": "^4.2.0",
-    "cucumber": "^6.0.5",
     "elliptic": "^6.5.3",
     "fabric-protos": "^2.2.2",
     "jsrsasign": "^9.1.3",
