+++ pkg/server/registry.go
