+++ scenario/fixtures/docker-compose/.env
@@ -1,3 +1,3 @@
 DOCKER_IMG_TAG=:2.2
 FABRIC_CA_TAG=:1.4
-THIRDPARTY_IMG_TAG=:0.4.16
+COUCHDB_IMG_TAG=:3.1
