+++ scenario/node/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@hyperledger/fabric-gateway-scenario",
+  "version": "0.0.1",
+  "description": "Scenario test for Fabric Gateway",
+  "engines": {
+    "node": "^10.13.0 || ^12.13.0"
+  },
+  "scripts": {
+    "test": "cucumber-js -f node_modules/cucumber-pretty --require 'lib/**/*.js' ../features"
+  },
+  "private": true,
+  "author": "",
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "cucumber": "^6.0.5",
+    "cucumber-pretty": "^6.0.0",
+    "fabric-gateway": "file:../../node"
+  }
+}
